# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "LaraSol API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/produces": {
            "get": {
                "tags": [
                    "Produces"
                ],
                "description": "Get all Produces",
                "operationId": "getProducesList",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Produces"
                ],
                "summary": "Store new Produce",
                "description": "Returns produce data",
                "operationId": "storeProduces",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreProduceRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Produces"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/produces/{id}": {
            "get": {
                "tags": [
                    "Produces"
                ],
                "summary": "Get produce information",
                "description": "Returns produce data",
                "operationId": "getProduceById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Produce id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Produces"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "put": {
                "tags": [
                    "Produces"
                ],
                "summary": "Update existing produce",
                "description": "Returns updated produce data",
                "operationId": "updateProduces",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Produce id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateProduceRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Produces"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Produces"
                ],
                "summary": "Delete existing produce",
                "description": "Deletes a record and returns no content",
                "operationId": "deleteProduce",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Produce id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StoreProduceRequest": {
                "title": "Store Produce Form request",
                "required": [
                    "id",
                    "produce_name"
                ],
                "properties": {
                    "produce_name": {
                        "description": "County Name",
                        "type": "string",
                        "example": "Veal"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "StoreProduceRequest"
                }
            },
...
            "UpdateProduceRequest": {
                "title": "Update Produce Form request",
                "required": [
                    "id",
                    "produce_name"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "produce_name": {
                        "description": "Name",
                        "type": "string"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "UpdateProduceRequest"
                }
            },
...
            "Produces": {
                "required": [
                    "produce_name"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "produce_name": {
                        "description": "Produces",
                        "type": "string",
                        "readOnly": "true"
                    },
                    "created_at": {
                        "$ref": "#/components/schemas/BaseModel/properties/created_at"
                    },
                    "updated_at": {
                        "$ref": "#/components/schemas/BaseModel/properties/updated_at"
                    },
                    "deleted_at": {
                        "$ref": "#/components/schemas/BaseModel/properties/deleted_at"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Produces"
                }
            },
...
        }
    }
}